GH-1509: split datasets into smaller submodules
